+++ packages/common/types/IllustrationNames.ts
@@ -1,11 +1,17 @@
 import type {
   HeroSquareName,
   PictogramName,
+  SpotIconName,
   SpotRectangleName,
   SpotSquareName,
 } from '@cbhq/cds-illustrations';
 
-export type IllustrationVariant = 'spotSquare' | 'spotRectangle' | 'heroSquare' | 'pictogram';
+export type IllustrationVariant =
+  | 'spotSquare'
+  | 'spotRectangle'
+  | 'heroSquare'
+  | 'pictogram'
+  | 'spotIcon';
 
 export type IllustrationHeroSquareNames = HeroSquareName;
 
@@ -14,3 +20,5 @@ export type IllustrationPictogramNames = PictogramName;
 export type IllustrationSpotRectangleNames = SpotRectangleName;
 
 export type IllustrationSpotSquareNames = SpotSquareName;
+
+export type IllustrationSpotIconNames = SpotIconName;