+++ packages/mobile/illustrations/SpotIcon.tsx
@@ -1,0 +1,9 @@
+import type { SpotIconProps as SpotIconCommonProps } from '@cbhq/cds-common/types/IllustrationProps';
+import svgJsMap from '@cbhq/cds-illustrations/__generated__/spotIcon/data/svgJsMap';
+
+import { createIllustration, IllustrationA11yProps } from './createIllustration';
+
+export const SpotIcon = createIllustration('spotIcon', svgJsMap);
+
+export type SpotIconProps = SpotIconCommonProps & IllustrationA11yProps;
+export type { SpotIconName } from '@cbhq/cds-illustrations';