+++ packages/web/illustrations/__stories__/Illustration.stories.tsx
@@ -11,3 +11,4 @@ export const spotSquare = getIllustrationSheet('spotSquare');
 export const spotRectangle = getIllustrationSheet('spotRectangle');
 export const heroSquare = getIllustrationSheet('heroSquare');
 export const pictogram = getIllustrationSheet('pictogram');
+export const spotIcon = getIllustrationSheet('spotIcon');