+++ packages/web/illustrations/__stories__/getIllustrationSheet.tsx
@@ -5,10 +5,12 @@ import type {
   IllustrationVariant,
   PictogramDimension,
   Spectrum,
+  SpotIconDimension,
 } from '@cbhq/cds-common';
 import { illustrationDimensions, illustrationSizes } from '@cbhq/cds-common/tokens/illustrations';
 import heroSquareVersionMap from '@cbhq/cds-illustrations/__generated__/heroSquare/data/versionMap';
 import pictogramVersionMap from '@cbhq/cds-illustrations/__generated__/pictogram/data/versionMap';
+import spotIconVersionMap from '@cbhq/cds-illustrations/__generated__/spotIcon/data/versionMap';
 import spotRectangleVersionMap from '@cbhq/cds-illustrations/__generated__/spotRectangle/data/versionMap';
 import spotSquareVersionMap from '@cbhq/cds-illustrations/__generated__/spotSquare/data/versionMap';
 
@@ -22,6 +24,8 @@ import {
   HeroSquareName,
   Pictogram,
   PictogramName,
+  SpotIcon,
+  SpotIconName,
   SpotRectangle,
   SpotRectangleName,
   SpotSquare,
@@ -49,6 +53,11 @@ const SIZES = {
     width: illustrationSizes[illustrationDimensions.heroSquare[0]][0],
     scaleMultiplier: 0.24,
   },
+  spotIcon: {
+    height: illustrationSizes[illustrationDimensions.spotIcon[0]][1],
+    width: illustrationSizes[illustrationDimensions.spotIcon[0]][0],
+    scaleMultiplier: 1,
+  },
 };
 
 function keys<T>(obj: { [key in keyof T]: T[key] }) {
@@ -60,6 +69,7 @@ const images = {
   pictogram: keys(pictogramVersionMap),
   spotRectangle: keys(spotRectangleVersionMap),
   spotSquare: keys(spotSquareVersionMap),
+  spotIcon: keys(spotIconVersionMap),
 };
 
 export function getIllustrationSheet<Type extends IllustrationVariant>(type: Type) {
@@ -117,10 +127,19 @@ export function getIllustrationSheet<Type extends IllustrationVariant>(type: Typ
             />
           )}
           {type === 'pictogram' && (
+            // render a 48x48 thumbnail
             <Pictogram
               name={name as PictogramName}
               dimension={dim as PictogramDimension}
               scaleMultiplier={SIZES.pictogram.scaleMultiplier}
+            />
+          )}
+          {type === 'spotIcon' && (
+            // render a 32x32 thumbnail
+            <SpotIcon
+              name={name as SpotIconName}
+              dimension={dim as SpotIconDimension}
+              scaleMultiplier={SIZES.spotIcon.scaleMultiplier}
             />
           )}
           <TextLegal as="p" noWrap>